Rework map editor
The editor states are now child classes of the abstract EditorState class.
The states are handled inside methods of these state classes to improve code readability.

Pathlines are now instances of PathLine class and are linked to waypoints directly as head and tail lines.
Because of this, it is easier to manage specific path lines via waypoints that are related.
For example, if a waypoint is moved, it is easier to update positions of the lines that are linked to that specific waypoint.