Implement texture unloading
Textures can now be tied to a specific scene. These scene specific textures can be unloaded with the unloadSceneSpecificTextures()-method.

If no scene is tied to the texture, the texture will be unloaded only if unloadCommonTextures() is called.

The benefit of this implementation is that a texture does not have to be unloaded and loaded again if it is used in multiple scenes.

However, in the end it is up to the scene to decide at which point the textures are loaded and unloaded.
Typically textures are loaded when the constructor of the scene is called and unloaded when the destructor is called.